requirements.txt
flake8
tox
isort
black
yamllint
openshift
docopt
